client cpp fix extended depth chunk type

+++ extensions/client_cpp/hl2ss_lnm.cpp
@@ -175,14 +175,13 @@ std::unique_ptr<hl2ss::rx_extended_audio> rx_extended_audio(char const* host, ui
     return decoded ? std::make_unique<hl2ss::rx_decoded_extended_audio>(host, port, chunk, mixer_mode, loopback_gain, microphone_gain, profile, level) : std::make_unique<hl2ss::rx_extended_audio>(host, port, chunk, mixer_mode, loopback_gain, microphone_gain, profile, level);
 }
 
-std::unique_ptr<hl2ss::rx_extended_depth> rx_extended_depth(char const* host, uint16_t port, uint64_t media_index, uint64_t stride_mask, uint16_t chunk, uint8_t mode, uint8_t divisor, uint8_t profile_z, bool decoded)
+std::unique_ptr<hl2ss::rx_extended_depth> rx_extended_depth(char const* host, uint16_t port, uint64_t media_index, uint64_t stride_mask, uint64_t chunk, uint8_t mode, uint8_t divisor, uint8_t profile_z, bool decoded)
 {
     std::vector<uint64_t> options;
     options.push_back(hl2ss::h26x_encoder_property::HL2SSAPI_VideoMediaIndex);
     options.push_back(media_index);
     options.push_back(hl2ss::h26x_encoder_property::HL2SSAPI_VideoStrideMask);
     options.push_back(stride_mask);
-
     return decoded ? std::make_unique<hl2ss::rx_decoded_extended_depth>(host, port, chunk, mode, divisor, profile_z, options) : std::make_unique<hl2ss::rx_extended_depth>(host, port, chunk, mode, divisor, profile_z, options);
 }
